Better prompts and bug fixes for update script

+++ update-sui.sh
@@ -11,7 +11,22 @@
 
 SUI_NODE_FOLDER="sui"
 SUI_NODE_SERVICE="sui-node"
+SUI_NODE_NETWORK="devnet"
 HOME_DIR="${HOME}"
+
+###################################################################################################
+# HELPER FUNCTIONS
+###################################################################################################
+
+## Make sure we are in the correct directory
+function ChangeDirectory() {
+  cd ~
+  eval "cd $HOME_DIR"
+}
+
+function exists() {
+  command -v "$1" >/dev/null 2>&1
+}
 
 ###################################################################################################
 # MAIN
@@ -20,7 +35,7 @@ HOME_DIR="${HOME}"
 echo "" && echo '[INFO] Checking script version'
 read -r current_version <VERSION
 remote_version=$(wget -qO- https://raw.githubusercontent.com/mrv777/sui-installation-scripts/main/VERSION)
-echo "[INFO] Current version: $current_version"
+echo "[INFO] Local version: $current_version"
 echo "[INFO] Remote version: $remote_version"
 if [ $remote_version -gt $current_version ]; then
   echo "" && echo "[ERROR] There is a new version of the script available. Please update."
@@ -30,25 +45,37 @@ fi
 echo "" && echo '[INFO] No script updates. Continuing...'
 
 echo "" && echo '[INFO] Checking node version'
-node_version=$(curl -sL https://api.github.com/repos/MystenLabs/sui/tags | jq -r ".[1].name" | cut -d "-" -f2)
-remote_node_version=$(grep 'version =' /${HOME_DIR}/${SUI_NODE_FOLDER}/crates/sui/Cargo.toml -m 1 | cut -d'"' -f 2)
-echo "[INFO] Current node version: $node_version"
+remote_node_version=$(curl -sL https://api.github.com/repos/MystenLabs/sui/tags | jq -r ".[1].name" | cut -d "-" -f2)
+node_version=$(grep 'version =' /${HOME_DIR}/${SUI_NODE_FOLDER}/crates/sui/Cargo.toml -m 1 | cut -d'"' -f 2)
+echo "[INFO] Local node version: $node_version"
 echo "[INFO] Remote node version: $remote_node_version"
 if [ $node_version == $remote_node_version ]; then
   echo "" && echo "[ERROR] There is no new version of the node software available."
   echo "Exiting..." && echo ""
   exit 2
 fi
-exit 0
 
-echo "" && echo '[INFO] Stopping service' && sleep 1
+[ "${RUN_UPDATE:-}" ] || read -r -p "Update available, would you like to update SUI? (Default yes): " RUN_UPDATE
+RUN_UPDATE=${RUN_UPDATE:-yes}
+if [ "$RUN_UPDATE" == "no" ]; then
+  echo "Exiting..." && echo ""
+  exit 1
+fi
+
+echo "" && echo '[INFO] Stopping service'
 sudo systemctl stop ${SUI_NODE_SERVICE}.service
+sleep 1
 
-echo "" && echo '[INFO] Removing old files' && sleep 1
-rm -rf /var/sui/db/* /var/sui/genesis.blob $HOME/sui
-source $HOME/.cargo/env
+echo "" && echo '[INFO] Removing old files'
+sudo rm -rf $HOME_DIR/sui
+source $HOME_DIR/.cargo/env
+sleep 1
 
-echo "" && echo '[INFO] Check out GIT repo' && sleep 1
+echo "" && echo "[INFO] Working in the directory: $HOME_DIR"
+ChangeDirectory
+sleep 1
+
+echo "" && echo '[INFO] Check out GIT repo'
 # Set up your fork of the Sui repository
 git clone https://github.com/MystenLabs/sui.git
 cd sui
@@ -64,17 +91,20 @@ cp crates/sui-config/data/fullnode-template.yaml fullnode.yaml
 curl -fLJO https://github.com/MystenLabs/sui-genesis/raw/main/${SUI_NODE_NETWORK}/genesis.blob
 # Set path of genesis file in config file
 sudo sed -i.bak "s|genesis-file-location:.*|genesis-file-location: \"${HOME_DIR}\/${SUI_NODE_FOLDER}\/genesis.blob\"|" /${HOME_DIR}/${SUI_NODE_FOLDER}/fullnode.yaml
+sleep 1
 
-echo "" && echo '[INFO] Build sui node' && sleep 1
+echo "" && echo '[INFO] Build sui node'
 cargo build --release -p sui-node
 sudo sed -i.bak 's/127.0.0.1/0.0.0.0/' fullnode.yaml
+sleep 1
 
-echo "" && echo '[INFO] Restarting service' && sleep 1
+echo "" && echo '[INFO] Restarting service'
 sudo systemctl restart ${SUI_NODE_SERVICE}.service
+sleep 1
 
 echo ""
 echo "==================================================="
-echo "" && echo '[INFO] Check Sui status' && sleep 1
+echo "" && echo '[INFO] Check Sui status'
 if [[ $(service sui-node status | grep active) =~ "running" ]]; then
   echo -e "Your Sui Node \e[32mis updated\e[39m!"
   echo -e "You can check node status by the command \e[7mservice sui-node status\e[0m"
